reddit-research-tool/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
openai>=1.12.0
pandas>=2.1.0
